+++ src/common_infrastructure/sample_library/sample_library_lua.cpp
@@ -413,6 +413,7 @@ struct TableFields<Region::AudioProperties> {
         GainDb,
         StartOffsetFrames,
         TuneCents,
+        FadeInFrames,
         Count,
     };
     static constexpr FieldInfo FieldInfo(Field f) {
@@ -455,6 +456,22 @@ struct TableFields<Region::AudioProperties> {
                     .set =
                         [](SET_FIELD_VALUE_ARGS) {
                             FIELD_OBJ.tune_cents = (f32)luaL_checknumber(ctx.lua, -1);
+                        },
+                };
+            case Field::FadeInFrames:
+                return {
+                    .name = "fade_in_frames",
+                    .description_sentence = "The number of frames to fade in the audio data.",
+                    .example = "0",
+                    .default_value = "0",
+                    .lua_type = LUA_TNUMBER,
+                    .required = false,
+                    .set =
+                        [](SET_FIELD_VALUE_ARGS) {
+                            auto const val = luaL_checkinteger(ctx.lua, -1);
+                            if (val < 0)
+                                luaL_error(ctx.lua, "'%s' should be a positive integer", info.name.data);
+                            FIELD_OBJ.fade_in_frames = (u32)val;
                         },
                 };
             case Field::Count: break;